Modify : PR 피드백 적용

+++ carts/views.py
@@ -2,8 +2,10 @@ import json
 
 from django.http  import JsonResponse
 from django.views import View
+from django.forms import ValidationError
 
 from carts.models    import Cart
+from carts.validator import validate_quantity
 from products.models import ProductColor, Image
 from cores.decorator import login_authorization
 
@@ -16,33 +18,30 @@ class CartView(View):
             product_id       = data['product_id']
             color_id         = data['color_id']
             quantity         = data['quantity']
-            product_color_id = ProductColor.objects.get(product_id = product_id, color_id = color_id)
-            inventory        = product_color_id.inventory
+            product_color    = ProductColor.objects.get(product_id = product_id, color_id = color_id)
+            inventory        = product_color.inventory
             
-            if quantity <= 0:
-                return JsonResponse({'message' : 'Quantity must be positive number'}, status=400)
+            validate_quantity(quantity, inventory)
             
-            if quantity > inventory:
-                return JsonResponse({'message' : 'Quantity cannot be more than inventory'}, status=400)
-            
-            if not Cart.objects.filter(
-                user = user,
-                product_color = product_color_id).exists():
-                Cart.objects.create(
-                    user          = user,
-                    product_color = product_color_id,
-                    quantity      = quantity
-                )
-                return JsonResponse({'message' : 'Cart created'}, status=201)
-            
-            cart           = Cart.objects.get(user = user, product_color = product_color_id)
+            cart, is_created  = Cart.objects.get_or_create(
+                user          = user,
+                product_color = product_color,
+                defaults      = {'quantity' : 0}
+            )
             cart.quantity += quantity
             cart.save()
-            return JsonResponse({'message' : 'Cart upgraded'}, status=200)
+            
+            if is_created:
+                return JsonResponse({'message' : 'Cart created'}, status=200)
+            else:
+                return JsonResponse({'message' : 'Cart updated'}, status=200)
         except ProductColor.DoesNotExist:
             return JsonResponse({'message' : 'Product does not exist'}, status=400)
         except KeyError:
             return JsonResponse({'message' : 'Key error'}, status=400)
+        except ValidationError as e:
+            return JsonResponse({'message' : e.message}, status=400)
+        
     
     @login_authorization
     def get(self, request):
@@ -50,7 +49,7 @@ class CartView(View):
         cart_data = Cart.objects.filter(user = user)            
         
         if not cart_data.exists():
-            return JsonResponse({'message' : 'Cart does not exist'}, status=401)
+            return JsonResponse({'result' : []}, status=404)
 
         result = [{
             'cart_id'      : cart.id,
@@ -67,50 +66,43 @@ class CartView(View):
         return JsonResponse({'result' : result}, status=200)
 
     @login_authorization
-    def patch(self, request):
+    def patch(self, request, cart_id):
         try:
             data      = json.loads(request.body)
             user      = request.user
-            cart_id   = request.GET.get('id')
             quantity  = data['quantity']
-            cart      = Cart.objects.get(id = int(cart_id))
+            cart      = Cart.objects.get(user = user, id = int(cart_id))
             inventory = cart.product_color.inventory
 
             if not cart_id:
                 return JsonResponse({'message' : 'Url path is wrong'}, status=400)
 
-            if quantity <= 0:
-                return JsonResponse({'message' : 'Quantity must be positive number'}, status=400)
+            validate_quantity(quantity, inventory)
             
-            if quantity > inventory:
-                return JsonResponse({'message' : 'Quantity cannot be more than inventory'}, status=400)
-            
-            cart          = Cart.objects.get(user = user, id = cart_id)
             cart.quantity = quantity
             cart.save()
 
-            return JsonResponse({
-                'message' : 'Quantity changed',
-                'quantity': cart.quantity
-                }, status=200)
+            return JsonResponse({'quantity': cart.quantity}, status=200)
         except Cart.DoesNotExist:
             return JsonResponse({'message' : 'Cart does not exist'}, status=400)  
         except KeyError:
             return JsonResponse({'message' : 'Key error'}, status=400)
+        except ValidationError as e:
+            return JsonResponse({'message' : e.message}, status=400)
 
     @login_authorization
     def delete(self, request):
-        try:
-            user = request.user
-            cart_id_list = request.GET.getlist('id')
-
-            if not cart_id_list:
-                return JsonResponse({'message' : 'Cart does not exist'}, status=400)
-
-            for cart_id in cart_id_list:
-                cart = Cart.objects.get(user = user, id = cart_id)
-                cart.delete()
-            return JsonResponse({'message' : 'Cart deleted'}, status=200)
-        except Cart.DoesNotExist:
+        user = request.user
+        cart_id_list = request.GET.getlist('id')
+        
+        if not cart_id_list:
             return JsonResponse({'message' : 'Cart does not exist'}, status=400)
-    
+        
+        cart = Cart.objects.filter(user = user, id__in = cart_id_list)
+        
+        if not cart:
+            return JsonResponse({'message' : 'Cart does not exist'}, status=400)
+        
+        cart.delete()
+        
+        return JsonResponse({'message' : 'Cart deleted'}, status=200)
